Split stocktake functionality out into separate file
